* Halo1_CE: Fixed a typo in CacheDefinitons.hpp where tag_versioning.project_yellow was checked twice, and tag_versioning.project_yellow_globals wasn't. Just so happened they're both currently at the same version! * Halo1_CE: Fixed a bad call to vsnprintf in PostProcessingErrorReporting.cpp. Was calc'ing the needed buffer size, should have been calling _vscprintf * Halo1_CE: Fixed typos in the object definitions where I used TStructGetPtrImpl when I should have been using TStructSubGetPtrImpl (as they're sub-structures)

+++ OpenSauce/Halo1/Halo1_CE/Memory/FunctionInterface.hpp
@@ -52,7 +52,7 @@ namespace Yelo
 				if(TBlockPreprocessCount == 0) return;
 
 				const function_process_proc* proc = &TBlockPreprocess[0];
-				for(int32 x = 0; x < TBlockPreprocessCount; proc++, x++)
+				for(size_t x = 0; x < TBlockPreprocessCount; proc++, x++)
 					if(proc != NULL)
 						(*proc)();
 			}
@@ -61,7 +61,7 @@ namespace Yelo
 				if(TBlockPostprocessCount == 0) return;
 
 				const function_process_proc* proc = &TBlockPostprocess[0];
-				for(int32 x = 0; x < TBlockPostprocessCount; proc++, x++)
+				for(size_t x = 0; x < TBlockPostprocessCount; proc++, x++)
 					if(proc != NULL)
 						(*proc)();
 			}
